+++ mandatory/includes/Solver.hpp
@@ -6,7 +6,7 @@
 /*   By: dcaetano <dcaetano@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/23 08:43:43 by dcaetano          #+#    #+#             */
-/*   Updated: 2026/02/23 10:06:55 by dcaetano         ###   ########.fr       */
+/*   Updated: 2026/02/25 17:51:18 by dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ class Solver
 {
 public:
 	friend std::ostream &operator<<(std::ostream &,
-									const std::map<long long int, double> &);
+									const std::map<unsigned long long int, long double> &);
 
 	static void solve(const Equation &);
 
@@ -28,15 +28,15 @@ private:
 	Solver &operator=(const Solver &);
 	~Solver();
 
-	static void __solveZeroDegreeEquation(const double &);
-	static void __solveFirstDegreeEquation(const double &,
-										   const double &);
-	static void __solveSecondDegreeEquation(const double &,
-											const double &,
-											const double &);
+	static void __solveZeroDegreeEquation(const long double &);
+	static void __solveFirstDegreeEquation(const long double &,
+										   const long double &);
+	static void __solveSecondDegreeEquation(const long double &,
+											const long double &,
+											const long double &);
 
-	static double __sqrt(const double &);
+	static long double __sqrt(const long double &);
 
-	static void __displayComplex(const double &,
-								 const double &);
+	static void __displayComplex(const long double &,
+								 const long double &);
 };
