+++ mandatory/includes/Term.hpp
@@ -6,7 +6,7 @@
 /*   By: dcaetano <dcaetano@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/23 08:01:58 by dcaetano          #+#    #+#             */
-/*   Updated: 2026/02/23 09:30:52 by dcaetano         ###   ########.fr       */
+/*   Updated: 2026/02/25 17:41:18 by dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,16 +18,16 @@ class Term
 {
 public:
 	Term(void);
-	Term(const double &,
-		 const long long int &);
+	Term(const long double &,
+		 const unsigned long long int &);
 	Term(const Term &);
 	Term &operator=(const Term &);
 	~Term();
 
-	const double &getValue(void) const;
-	const long long int &getDegree(void) const;
+	const long double &getValue(void) const;
+	const unsigned long long int &getDegree(void) const;
 
 private:
-	double __value;
-	long long int __degree;
+	long double __value;
+	unsigned long long int __degree;
 };
