+++ mandatory/srcs/Solver.cpp
@@ -6,7 +6,7 @@
 /*   By: dcaetano <dcaetano@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/23 08:45:50 by dcaetano          #+#    #+#             */
-/*   Updated: 2026/02/24 13:06:26 by dcaetano         ###   ########.fr       */
+/*   Updated: 2026/02/25 18:03:03 by dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@ Solver &Solver::operator=(const Solver &) { return *this; }
 Solver::~Solver() {}
 
 std::ostream &operator<<(std::ostream &os,
-						 const std::map<long long int, double> &p_reduced)
+						 const std::map<unsigned long long int, long double> &p_reduced)
 {
 	if (p_reduced.empty() == true)
 		return os << "0 = 0";
-	for (std::map<long long int, double>::const_iterator t = p_reduced.begin(); t != p_reduced.end(); t++)
+	for (std::map<unsigned long long int, long double>::const_iterator t = p_reduced.begin(); t != p_reduced.end(); t++)
 	{
 		if (t != p_reduced.begin())
 		{
@@ -41,14 +41,14 @@ std::ostream &operator<<(std::ostream &os,
 
 void Solver::solve(const Equation &p_eq)
 {
-	std::map<long long int, double> reducedForm;
+	std::map<unsigned long long int, long double> reducedForm;
 	for (std::vector<Term>::const_iterator t = p_eq.__leftSide.begin(); t != p_eq.__leftSide.end(); t++)
 		reducedForm[t->getDegree()] += t->getValue();
 	for (std::vector<Term>::const_iterator t = p_eq.__rightSide.begin(); t != p_eq.__rightSide.end(); t++)
 		reducedForm[t->getDegree()] -= t->getValue();
 	std::cout << "Reduced form: " << reducedForm << std::endl;
-	long long int polynomialDegree = 0;
-	for (std::map<long long int, double>::const_iterator t = reducedForm.begin(); t != reducedForm.end(); t++)
+	unsigned long long int polynomialDegree = 0;
+	for (std::map<unsigned long long int, long double>::const_iterator t = reducedForm.begin(); t != reducedForm.end(); t++)
 		polynomialDegree = std::max(polynomialDegree, t->first);
 	if (polynomialDegree == 0)
 		return Solver::__solveZeroDegreeEquation(reducedForm[polynomialDegree]);
@@ -60,10 +60,10 @@ void Solver::solve(const Equation &p_eq)
 	std::cout << "The polynomial degree is strictly greater than 2, I can't solve." << std::endl;
 }
 
-void Solver::__solveZeroDegreeEquation(const double &a) { std::cout << (a == 0 ? "Any real number is a solution." : "No solution.") << std::endl; }
+void Solver::__solveZeroDegreeEquation(const long double &a) { std::cout << (a == 0 ? "Any real number is a solution." : "No solution.") << std::endl; }
 
-void Solver::__solveFirstDegreeEquation(const double &a,
-										const double &b)
+void Solver::__solveFirstDegreeEquation(const long double &a,
+										const long double &b)
 {
 	if (a == 0)
 		return Solver::__solveZeroDegreeEquation(b);
@@ -71,13 +71,13 @@ void Solver::__solveFirstDegreeEquation(const double &a,
 	std::cout << (b == 0.0 ? 0.0 : -b / a) << std::endl;
 }
 
-void Solver::__solveSecondDegreeEquation(const double &a,
-										 const double &b,
-										 const double &c)
+void Solver::__solveSecondDegreeEquation(const long double &a,
+										 const long double &b,
+										 const long double &c)
 {
 	if (a == 0.0)
 		return Solver::__solveFirstDegreeEquation(b, c);
-	double disc = b * b - 4.0 * a * c;
+	long double disc = b * b - 4.0 * a * c;
 	if (disc > 0)
 	{
 		std::cout << "Discriminant is strictly positive, the two solutions are:" << std::endl;
@@ -97,14 +97,14 @@ void Solver::__solveSecondDegreeEquation(const double &a,
 	std::cout << (b == 0.0 ? 0.0 : -b / (2.0 * a)) << std::endl;
 }
 
-double Solver::__sqrt(const double &x)
+long double Solver::__sqrt(const long double &x)
 {
 	if (x < 0.0)
-		return -std::numeric_limits<double>::quiet_NaN();
+		return -std::numeric_limits<long double>::quiet_NaN();
 	if (x == 0.0 || x == 1.0)
 		return x;
-	double approx = x;
-	double next;
+	long double approx = x;
+	long double next;
 	while (true)
 	{
 		next = 0.5 * (approx + x / approx);
@@ -115,8 +115,8 @@ double Solver::__sqrt(const double &x)
 	return approx;
 }
 
-void Solver::__displayComplex(const double &realPart,
-							  const double &imagPart)
+void Solver::__displayComplex(const long double &realPart,
+							  const long double &imagPart)
 {
 	if (realPart == 0.0 && imagPart == 0.0)
 	{
