+++ mandatory/srcs/Term.cpp
@@ -6,7 +6,7 @@
 /*   By: dcaetano <dcaetano@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/23 08:04:15 by dcaetano          #+#    #+#             */
-/*   Updated: 2026/02/23 11:17:12 by dcaetano         ###   ########.fr       */
+/*   Updated: 2026/02/25 17:41:29 by dcaetano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 Term::Term(void) : __value(0.0),
 				   __degree(0.0) {}
 
-Term::Term(const double &value,
-		   const long long int &degree) : __value(value),
-										  __degree(degree) {}
+Term::Term(const long double &value,
+		   const unsigned long long int &degree) : __value(value),
+												   __degree(degree) {}
 
 Term::Term(const Term &copy) : __value(copy.__value),
 							   __degree(copy.__degree) {}
@@ -34,6 +34,6 @@ Term &Term::operator=(const Term &other)
 
 Term::~Term() {}
 
-const double &Term::getValue(void) const { return this->__value; }
+const long double &Term::getValue(void) const { return this->__value; }
 
-const long long int &Term::getDegree(void) const { return this->__degree; }
+const unsigned long long int &Term::getDegree(void) const { return this->__degree; }
